+++ royalflush/_agent/observer.py
@@ -1,12 +1,10 @@
-from typing import TYPE_CHECKING
+from typing import Optional
 
 from aioxmpp import JID
 from spade.template import Template
 
-from ..agent import AgentBase
-
-if TYPE_CHECKING:
-    from ..behaviour.observer import ObserverBehaviour
+from .._agent.base import AgentBase
+from .._behaviour.observer import ObserverBehaviour
 
 
 class ObserverAgent(AgentBase):
@@ -21,7 +19,7 @@ class ObserverAgent(AgentBase):
         verify_security: bool = False,
     ):
         self.agents_observed: list[JID] = []
-        self.observation_theme_behaviours: dict[str, None | "ObserverBehaviour"] = {
+        self.observation_theme_behaviours: dict[str, Optional[ObserverBehaviour]] = {
             "message": None,
             "nn": None,
             "iteration": None,
@@ -36,7 +34,6 @@ class ObserverAgent(AgentBase):
         )
 
     async def setup(self) -> None:
-        from ..behaviour.observer import ObserverBehaviour
 
         await super().setup()
         for theme in self.observation_theme_behaviours: