+++ royalflush/_behaviour/premiofl/layer_receiver.py
@@ -9,7 +9,7 @@ from ...datatypes.consensus import Consensus
 from ...message.message import RfMessage
 
 if TYPE_CHECKING:
-    from ...agent.base import PremioFlAgent
+    from ..._agent.base import PremioFlAgent
 
 
 class LayerReceiverBehaviour(CyclicBehaviour):