+++ royalflush/agent/premiofl/__init__.py
@@ -1,5 +1,5 @@
-from .acol import AcolAgent
-from .macofl import MacoflAgent
-from .pmacofl_min import PmacoflMinAgent
+from ..._agent.premiofl.acol import AcolAgent
+from ..._agent.premiofl.macofl import MacoflAgent
+from ..._agent.premiofl.pmacofl_min import PmacoflMinAgent
 
 __all__ = ["AcolAgent", "MacoflAgent", "PmacoflMinAgent"]