+++ royalflush/behaviour/premiofl/__init__.py
@@ -1,3 +1,0 @@
-from .layer_receiver import LayerReceiverBehaviour
-
-__all__ = ["LayerReceiverBehaviour"]